+++ src/client/public/locale/de/common.json
@@ -198,6 +198,7 @@
   "header_settings": "Einstellungen",
   "height_reference_system": "Höhenreferenzsystem",
   "hidden": "Versteckt",
+  "hideTable": "Tabelle verbergen",
   "hour": "Stunde",
   "hours": "Stunden",
   "humidity": "Feuchtigkeit",
@@ -348,6 +349,7 @@
   "showAll": "Alle anzeigen",
   "showallfields": "Alle Felder anzeigen",
   "showHistory": "Verlauf anzeigen",
+  "showTable": "Tabelle anzeigen",
   "someSelected": "{{howMany}} Bohrungen ausgewählt",
   "sourceCode": "Quellcode",
   "spatial_reference_system": "Räumliches Bezugssystem",
