+++ src/client/public/locale/en/common.json
@@ -198,6 +198,7 @@
   "header_settings": "Settings",
   "height_reference_system": "Height reference system",
   "hidden": "Hidden",
+  "hideTable": "Hide table",
   "hour": "hour",
   "hours": "hours",
   "humidity": "Humidity",
@@ -348,6 +349,7 @@
   "showAll": "Show all",
   "showallfields": "Show all fields",
   "showHistory": "Show history",
+  "showTable": "Show table",
   "someSelected": "{{howMany}} boreholes selected",
   "sourceCode": "Source code",
   "spatial_reference_system": "Spatial reference system",
