+++ src/client/public/locale/fr/common.json
@@ -198,6 +198,7 @@
   "header_settings": "Paramètres",
   "height_reference_system": "Système altitudinal de référence",
   "hidden": "Masqué",
+  "hideTable": "Cacher le tableau",
   "hour": "heure",
   "hours": "heures",
   "humidity": "Humidité",
@@ -348,6 +349,7 @@
   "showAll": "Afficher tout",
   "showallfields": "Afficher tous les champs",
   "showHistory": "Montrer l'historique",
+  "showTable": "Afficher le tableau",
   "someSelected": "{{howMany}} forages sélectionnés",
   "sourceCode": "Code source",
   "spatial_reference_system": "Système de cordonnées",
