+++ src/client/public/locale/it/common.json
@@ -198,6 +198,7 @@
   "header_settings": "Impostazioni",
   "height_reference_system": "Sistema di riferimento altitudinale",
   "hidden": "Nascosto",
+  "hideTable": "Nascondere la tabella",
   "hour": "ora",
   "hours": "ore",
   "humidity": "Umidità",
@@ -348,6 +349,7 @@
   "showAll": "Mostra tutti",
   "showallfields": "Mostra tutti i campi",
   "showHistory": "Mostra la storia",
+  "showTable": "Mostra la tabella",
   "someSelected": "{{howMany}} perforazioni selezionate",
   "sourceCode": "Codice sorgente",
   "spatial_reference_system": "Sistema di coordinate",
